Refactoring. Changes printers folders.
