+++ Task_02/Printers/Abstract/NumberOfMatches/NumberOfMatchesPrinter.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using Task_02.Statistics;
 
-namespace Task_02.Printers
+namespace Task_02.Printers.Abstract.NumberOfMatches
 {
     internal abstract class NumberOfMatchesPrinter : IPrinter
     {
