+++ Task_02/Printers/Abstract/WinRate/WinRatePrinter.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using Task_02.Statistics;
 
-namespace Task_02.Printers
+namespace Task_02.Printers.Abstract.WinRate
 {
     internal abstract class WinRatePrinter : IPrinter
     {
