+++ Task_02/Printers/Abstract/WinStrickPrinter.cs
