+++ Task_02/Program.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-using Task_02.Printers;
+using Task_02.Printers.Certain;
 using Task_02.Statistics;
 
 namespace Task_02
