+++ Task_02/Statistics/HeroUsagesStatisticMostFavoriteProvider.cs
@@ -1,11 +1,11 @@
 ﻿using System.Collections.Generic;
-using Task_02.Printers;
+using Task_02.Printers.Abstract.NumberOfMatches;
 
 namespace Task_02.Statistics
 {
     internal class HeroUsagesStatisticMostFavoriteProvider : AbstractStatisticProvider
     {
-        public HeroUsagesStatisticMostFavoriteProvider(List<HeroScoreInfo> infos) : base (infos, new FavoritePrinter()) { }
+        public HeroUsagesStatisticMostFavoriteProvider(List<HeroScoreInfo> infos) : base(infos, new FavoritePrinter()) { }
 
         internal override void Show()
         {
