+++ Task_02/Statistics/HeroUsagesStatisticMostSuccesfulProvider.cs
@@ -1,5 +1,5 @@
 ﻿using System.Collections.Generic;
-using Task_02.Printers;
+using Task_02.Printers.Abstract.WinRate;
 
 namespace Task_02.Statistics
 {
