+++ Task_02/Statistics/HeroUsagesStatisticMostUnfavoriteProvider.cs
@@ -1,5 +1,5 @@
 ﻿using System.Collections.Generic;
-using Task_02.Printers;
+using Task_02.Printers.Abstract.NumberOfMatches;
 
 namespace Task_02.Statistics
 {
